Extract summary functions
